Move /comms around
