+++ src/index.ts
@@ -3,7 +3,7 @@ require('dotenv').config()
 import express from 'express'
 import { Bot } from './lib/bot'
 import functions from './functions'
-import { VoiceChannel } from 'discord.js'
+import createCommsRouter from './lib/comms'
 
 const port = process.env.WEBSITES_PORT || process.env.PORT || 8080
 
@@ -35,39 +35,7 @@ app.get('/', (req, res) => {
   res.send(`<h1>${susGen()} is sus.</h1>`)
 })
 
-app.post('/comms', async (req, res) => {
-  const { voiceChannelId, usersThatCanSpeak } = req.body
-
-  // Look up the voice channel by ID
-  const voiceChannel = (await bot.channels.fetch(
-    voiceChannelId
-  )) as VoiceChannel
-  if (voiceChannel.type !== 'voice') return res.sendStatus(418)
-
-  await Promise.all(
-    voiceChannel.members.map(async (member) => {
-      // TODO: Make this real smart g00d
-      const discordMemberCanSpeak = usersThatCanSpeak.includes(
-        member.user.username
-      )
-
-      if (discordMemberCanSpeak) {
-        const permission = voiceChannel.permissionOverwrites.find((value) => {
-          return value.type === 'member' && value.id === member.user.id
-        })
-
-        await permission?.delete()
-      } else {
-        await voiceChannel.createOverwrite(member.user, { SPEAK: false })
-      }
-
-      await member.voice.setChannel(voiceChannel)
-    })
-  )
-
-  // Mute (but not server mute) everyone
-  return res.sendStatus(200)
-})
+app.use('/comms', createCommsRouter(bot))
 
 app.listen(port, () => {
   bot.log.success(`Server running at port ${port}`)
